clean pipeline
